DiscordVoiceBot/DiscordVoiceBot/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4.0",
    "ffmpeg-python>=0.2.0",
    "flask>=3.1.0",
    "gtts>=2.5.4",
    "pynacl>=1.5.0",
]
